Fix Warning

+++ JazzInParis/JazzInParis/MGBox/Categories/NSAttributedString+MGTrim.m
@@ -15,10 +15,10 @@
   unsigned loc, len;
 
   NSRange range = [string rangeOfCharacterFromSet:invertedSet];
-  loc = (range.length > 0) ? range.location : 0;
+  loc = (range.length > 0) ? (int)range.location : 0;
 
   range = [string rangeOfCharacterFromSet:invertedSet options:NSBackwardsSearch];
-  len = (range.length > 0) ? NSMaxRange(range) - loc : string.length - loc;
+  len = (range.length > 0) ? (int)NSMaxRange(range) - loc : (int)string.length - (int)loc;
 
   return [self attributedSubstringFromRange:NSMakeRange(loc, len)];
 }
